Extract duplicate code to shared folder
* add lib/shared to keep shared functions
* remove commented out code in the shooter prototype
  since the responsibility of collision detection will
  be moved to the bubbles themselves
