+++ lib/bubble.js
@@ -1,3 +1,6 @@
+const GetRandomColor = require('./shared/bubble-colors');
+const DrawBubble     = require('./shared/draw-bubbles');
+
 function Bubble(x, y, r, canvas, context) {
   this.x       = x;
   this.y       = y;
@@ -8,17 +11,12 @@ function Bubble(x, y, r, canvas, context) {
   this.oRadius = 10;
   this.canvas  = canvas;
   this.context = context;
-  this.color   = getRandomColor(this, context);
+  this.color   = GetRandomColor(this, context);
 }
 
 Bubble.prototype = {
   draw: function() {
-    this.context.beginPath();
-    this.context.arc(this.x, this.y, this.radius, this.sAngle, this.eAngle);
-
-    this.context.fillStyle = this.color;
-    this.context.fill();
-
+    DrawBubble(this);
     return this;
   },
 
@@ -27,15 +25,5 @@ Bubble.prototype = {
     return this;
   }
 };
-
-function getRandomColor(self, context) {
-  var colors = ['deeppink', 'navy', 'mediumslateblue', 'darkgreen', 'darkorange']
-
-  var gradient = context.createRadialGradient(self.x, self.y, self.iRadius, self.x, self.y, self.oRadius);
-  gradient.addColorStop(0, 'ivory');
-  gradient.addColorStop(1, colors[Math.floor(Math.random()*colors.length)]);
-
-  return gradient;
-}
 
 module.exports = Bubble;
