+++ lib/index.js
@@ -38,6 +38,7 @@ function onKeyUp (e) {
 requestAnimationFrame(function gameLoop() {
   context.clearRect(0, 0, canvas.width, canvas.height);
   window.addEventListener('keypress', onKeyUp, false);
+
   drawBubbles();
   drawShooter();
