+++ lib/shooter.js
@@ -1,4 +1,6 @@
-const Bubble = require('./bubble');
+const Bubble         = require('./bubble');
+const GetRandomColor = require('./shared/bubble-colors');
+const DrawBubble     = require('./shared/draw-bubbles');
 
 function Shooter(x, y, r, canvas, context) {
   this.x       = x;
@@ -10,17 +12,12 @@ function Shooter(x, y, r, canvas, context) {
   this.oRadius = 10;
   this.canvas  = canvas;
   this.context = context;
-  this.color   = getRandomColor(this, context);
+  this.color   = GetRandomColor(this, context);
 }
 
 Shooter.prototype = {
   draw: function() {
-    this.context.beginPath();
-    this.context.arc(this.x, this.y, this.radius, this.sAngle, this.eAngle);
-
-    this.context.fillStyle = this.color;
-    this.context.fill();
-
+    DrawBubble(this);
     return this;
   },
 
@@ -31,10 +28,6 @@ Shooter.prototype = {
   checkBackWall: function() {
     return this.y < 440 ? true : false;
   },
-
-  // checkForAdjacentBubbles: function () {
-  //   this.x === bubble.x - 11 || this.x === bubble.x + 11 || this.y === bubble.y - 11 ? true : false;
-  // },
 
   moveLeft: function() {
     if (this.checkInBounds()) {
@@ -51,7 +44,6 @@ Shooter.prototype = {
     } else if (!this.checkInBounds() && this.x === 10) {
       this.x = this.x + 5
     }
-    // this.checkInBounds() ? this.x = this.x + 5 : this;
     return this;
   },
 
@@ -62,16 +54,6 @@ Shooter.prototype = {
     this.checkBackWall() ? this.y = this.y + 1 : this;
     return this;
   }
-}
-
-function getRandomColor(self, context) {
-  var colors = ['deeppink', 'navy', 'mediumslateblue', 'darkgreen', 'darkorange']
-
-  var gradient = context.createRadialGradient(self.x, self.y, self.iRadius, self.x, self.y, self.oRadius);
-  gradient.addColorStop(0, 'ivory');
-  gradient.addColorStop(1, colors[Math.floor(Math.random()*colors.length)]);
-
-  return gradient;
 }
 
 module.exports = Shooter;
